priv/python/requirements.txt
amazon.ion==0.5.0
ionhash==1.1.0
six==1.13.0
